chore: move logs to internal
